moved components to new component sub folder
