+++ stagewhisper/app/src/components/transcription/Transcription.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+// Component for displaying transcription progress / results
+function Transcription() {
+  return <div>Transcription</div>;
+}
+
+export default Transcription;
